# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a FizzBuzz API.",
        "title": "FizzBuzz API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/fizzbuzz": {
            "post": {
                "description": "Generate a FizzBuzz sequence based on the provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Generate a FizzBuzz sequence",
                "operationId": "fizzbuzz-handler",
                "parameters": [
                    {
                        "description": "FizzBuzz parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.Request": {
            "type": "object",
            "properties": {
                "int1": {
                    "type": "integer"
                },
                "int2": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "string1": {
                    "type": "string"
                },
                "string2": {
                    "type": "string"
                }
            }
        }
    }
}